README should mention need to escape #chan name
Most/all UNIX-y shells use the octhorpe (#) for comments. As such,
if the home channel name supplied on the command line isn't escaped,
it will neve be passed to the binary.

Escaping can either be done by encasing the channel name in quotes or
by escaping the octhorpe alone with a backslash.